account for size change

+++ plugins/devel/counters.cpp
@@ -25,7 +25,7 @@ command_result df_counters (color_ostream &out, vector <string> & parameters)
     for (size_t i = 0; i < counters.size(); i++)
     {
         auto counter = counters[i];
-        out.print("%i (%s): %i\n", counter->id, ENUM_KEY_STR(misc_trait_type, counter->id).c_str(), counter->value);
+        out.print("%i (%s): %i\n", (int)counter->id, ENUM_KEY_STR(misc_trait_type, counter->id).c_str(), counter->value);
     }
 
     return CR_OK;
